scfd-protein-folding/requirements.txt
# Core dependencies for AlphaFold-SCFD workflow
numpy>=1.21.0
biopython>=1.79
tqdm>=4.62.0

# Visualization dependencies
matplotlib>=3.5.0
plotly>=5.0.0

# Optional but recommended
scipy>=1.7.0
pandas>=1.3.0

# For Jupyter notebook integration (optional)
jupyter>=1.0.0
ipywidgets>=7.6.0